fix file without extension

+++ distributed/utils.py
@@ -1018,24 +1018,25 @@ def import_file(path):
     names_to_import = []
     tmp_python_path = None
 
-    if ext in ('.py'):  # , '.pyc'):
-        if directory not in sys.path:
-            tmp_python_path = directory
-        names_to_import.append(name)
-    if ext == '.py':  # Ensure that no pyc file will be reused
-        cache_file = cache_from_source(path)
-        with ignoring(OSError):
-            os.remove(cache_file)
-    if ext in ('.egg', '.zip', '.pyz'):
-        if path not in sys.path:
-            sys.path.insert(0, path)
-        if ext == '.egg':
-            import pkg_resources
-            pkgs = pkg_resources.find_distributions(path)
-            for pkg in pkgs:
-                names_to_import.append(pkg.project_name)
-        elif ext in ('.zip', '.pyz'):
+    if ext.strip() != '':
+        if ext in ('.py'):  # , '.pyc'):
+            if directory not in sys.path:
+                tmp_python_path = directory
             names_to_import.append(name)
+        if ext == '.py':  # Ensure that no pyc file will be reused
+            cache_file = cache_from_source(path)
+            with ignoring(OSError):
+                os.remove(cache_file)
+        if ext in ('.egg', '.zip', '.pyz'):
+            if path not in sys.path:
+                sys.path.insert(0, path)
+            if ext == '.egg':
+                import pkg_resources
+                pkgs = pkg_resources.find_distributions(path)
+                for pkg in pkgs:
+                    names_to_import.append(pkg.project_name)
+            elif ext in ('.zip', '.pyz'):
+                names_to_import.append(name)
 
     loaded = []
     if not names_to_import:
